# internal/routes/docs/swagger.yaml
definitions:
  database.Transaction:
    properties:
      amount:
        type: integer
      date:
        type: string
      id:
        type: string
      platform_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  routes.UpdateTrade.request:
    properties:
      platform_id:
        type: string
      status:
        type: string
    required:
    - platform_id
    - status
    type: object
info:
  contact: {}
paths:
  /trades/{id}:
    get:
      description: Get a specific trade by its ID
      parameters:
      - description: Trade ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Transaction'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get trade by ID
      tags:
      - trades
    post:
      consumes:
      - application/json
      description: Update the status of a trade by ID and platform ID
      parameters:
      - description: Trade ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateTrade.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update trade status
      tags:
      - trades
  /trades/platform/{id}:
    get:
      description: Get a list of all trades
      parameters:
      - description: Platform ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all trades by platform id
      tags:
      - trades
swagger: "2.0"